requirements.txt
django>=4.2 
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
psycopg2-binary>=2.9
Pillow
gunicorn==21.2.0

